# backend/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Plataforma MOOC — API
  version: 0.2.0
  description: |
    Plataforma web de cursos masivos abiertos en línea.

    ## Convenciones transversales

    - **Autenticación.** Token opaco en `Authorization: Bearer <token>`. La
      sesión se resuelve contra Redis en cada petición, así que revocarla surte
      efecto de inmediato (ADR-0006).
    - **Errores.** Siempre `application/problem+json` (RFC 9457), con
      `trace_id` para correlacionar con logs y trazas, y un arreglo `errors`
      que lista **todos** los incumplimientos, no el primero (ADR-0009).
    - **Idempotencia.** Los endpoints marcados exigen `Idempotency-Key`.
      Repetir la petición con la misma clave y el mismo cuerpo devuelve la
      respuesta guardada con `Idempotency-Replayed: true`; con cuerpo distinto,
      `422` (ADR-0008).
    - **Concurrencia.** Los recursos editables devuelven `ETag`, y su `PATCH`
      exige `If-Match`. Si no coincide, `412` con el estado del servidor.
    - **Paginación.** Siempre por cursor opaco, nunca por desplazamiento.
    - **Acciones.** Van como segmento final de la ruta
      (`POST .../versions/1/publish`), no con dos puntos. El `ServeMux` de la
      stdlib exige que un comodín ocupe el segmento entero, y elegimos la
      stdlib en el ADR-0002.
    - **Binarios.** No pasan por la API: se suben y descargan con credenciales
      firmadas contra el almacenamiento de objetos (ADR-0005).

    ## Invariantes que este contrato hace cumplir

...
servers:
...
security:
  - sesion: []
paths:
...
  /api/v1/auth/register:
    post:
      tags: [identidad]
      operationId: registerStudent
      summary: Registro público de estudiante
      description: |
        Solo crea estudiantes: los profesores se crean por administración.

        Responde `202` sin esperar al worker que envía el correo. El trabajo se
        registra en PostgreSQL dentro de la misma transacción que el usuario y
        se publica en la cola después del commit, de modo que un fallo de Redis
        no pierde el correo (ADR-0004).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "202":
          description: Cuenta creada, pendiente de verificación
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegisterResponse" }
        "409": { $ref: "#/components/responses/Conflicto" }
        "422": { $ref: "#/components/responses/Validacion" }
        "429": { $ref: "#/components/responses/DemasiadasSolicitudes" }
  /api/v1/auth/verify-email:
    post:
      tags: [identidad]
      operationId: verifyEmail
      summary: Verificar el correo con el token recibido
      description: El token es de un solo uso y expira a las 24 horas.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, minLength: 1 }
      responses:
        "200":
          description: Correo verificado
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified: { type: boolean, const: true }
                  detail: { type: string }
        "422": { $ref: "#/components/responses/Validacion" }
...
  /api/v1/auth/login:
    post:
      tags: [identidad]
      operationId: login
      summary: Iniciar sesión
      description: |
        Devuelve un token opaco de sesión.

        El mismo `401` y el mismo mensaje para usuario inexistente, contraseña
        incorrecta, cuenta sin verificar y cuenta suspendida: no se revela qué
        correos existen ni en qué estado están. Limitado a 5 por minuto e IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Sesión abierta
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "422": { $ref: "#/components/responses/Validacion" }
        "429": { $ref: "#/components/responses/DemasiadasSolicitudes" }
  /api/v1/auth/logout:
    post:
      tags: [identidad]
      operationId: logout
      summary: Cerrar la sesión actual
      description: |
        Borra la sesión de Redis. La petición siguiente con ese token responde
        `401`, sin ventana de gracia.
      responses:
        "204": { description: Sesión revocada }
        "401": { $ref: "#/components/responses/NoAutenticado" }
...
  /api/v1/me:
    get:
      tags: [identidad]
      operationId: getCurrentUser
      summary: Perfil de la sesión actual
      responses:
        "200":
          description: Perfil
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
    patch:
      tags: [identidad]
      operationId: updateCurrentUser
      x-estado: planificado
      summary: Actualizar el perfil propio
      description: En la Entrega 1 solo el nombre; el correo no cambia.
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name: { type: string, minLength: 1, maxLength: 120 }
      responses:
        "200":
          description: Perfil actualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/NoAutenticado" }
        "412": { $ref: "#/components/responses/PrecondicionFallida" }
        "422": { $ref: "#/components/responses/Validacion" }
...
components:
  securitySchemes:
    sesion:
      type: http
      scheme: bearer
      description: |
        Token opaco de 32 bytes emitido por `/api/v1/auth/login`. No es un JWT:
        no lleva información dentro y se puede revocar (ADR-0006).
  headers:
    ETag:
      description: Versión del recurso. Se devuelve en `If-Match` al modificarlo.
      schema: { type: string }
...
    RetryAfter:
      description: Segundos que conviene esperar antes de reintentar.
      schema: { type: integer }
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: "`ETag` obtenido al leer el recurso. Si no coincide, `412`."
      schema: { type: string }
...
  responses:
    NoAutenticado:
      description: Sin sesión, o la sesión fue revocada o expiró
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Conflicto:
      description: Estado incompatible con la operación
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    PrecondicionFallida:
      description: |
        El `If-Match` no coincide: alguien modificó el recurso mientras tanto.
        El cuerpo trae el estado actual del servidor.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Validacion:
      description: |
        La solicitud tiene campos inválidos. `errors` lista **todos** los
        incumplimientos.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    DemasiadasSolicitudes:
      description: Límite de tasa superado
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Problem:
      type: object
      description: Error uniforme, RFC 9457.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, examples: ["https://mooc.local/errors/validation_failed"] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        trace_id:
          type: string
          description: Correlaciona esta respuesta con los logs y las trazas.
        errors:
          type: array
          description: |
            **Todos** los incumplimientos, no el primero. Es la mecánica que
            exige CA-01 para la publicación de un curso.
          items: { $ref: "#/components/schemas/FieldError" }
    FieldError:
      type: object
      required: [code]
      properties:
        code: { type: string, examples: ["structure.minimum_not_met"] }
        field: { type: string, examples: ["modules[0].units[1].resources[0]"] }
        detail: { type: string }
...
    RegisterRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, format: password, minLength: 10, maxLength: 128 }
        full_name: { type: string, minLength: 1, maxLength: 120 }
    RegisterResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        status: { type: string, const: pending_verification }
        detail: { type: string }
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: Token opaco de sesión. Va en la cabecera de autorización.
        expires_at: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string }
        role: { type: string, enum: [admin, teacher, student] }
        status: { type: string, enum: [active, suspended, deleted] }
        email_verified_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
...